kms_cursor_legacy: Make tests less strict.
Some tests may wait a vblank on master. Instead of
failing the tests, allow them to succeed if 1 vblank
passed.

Also no longer warn if a page flip is missed, change it to
info instead. As long as no more than 25% of the vblanks are
missed the tests will pass.

Also fix nonblocking modeset vs cursor, which needs an array
of cursor argument. This was a nice buffer overrun before.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>