kms_frontbuffer_tracking: fix sink CRC assertion
If we already detected an error, don't try to assert the size of what
we didn't read.

In machines where the sink CRC is unreliable, this was failing tests
where the sink CRC is not mandatory (FBC tests). With this change we
won't fail anymore, we'll just print error messages saying that the
sink CRC is unreliable.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>