kms_cursor_legacy: Run cursor updates only for a quarter of the time.
Because in the ideal case we currently wait 50% of the time for every frame,
on some machines cpu scaling gets in the way when we idle a lot and will
cause the machine to throttle to a lower cpu speed. This causes a
failure because we may end up missing vblanks.

Work around this by only running for 1/4th of the time at max speed,
even on low cpu speeds it will be less likely to run into issues then.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>