kms_cursor_legacy: Use cursor movement instead of hiding for 2x tests
This forces the cursor fastpath to be used. While at it, add -atomic
versions of the test to compare legacy pageflip vs atomic.

The nonblocking modeset test is changed slightly, due to cdclk changes
the page flip may be stalled so it's not useful to test it. Just test
that a queued page flip + nonblocking modeset works as intended and
there is no kernel spew.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>