kms_busy: Add extended tests that force a hang.
The intel driver adds a timeout to each atomic update of 10s for
the new fb.

Make sure this works as intended, and also test what happens when
hanging the old framebuffer, with and without modeset. The driver
waits indefinitely on the old framebuffer, but another blocking
update will wait at most 10 seconds for it.

Abuse this fact in a testcase. For added bonus also add some tests
that make sure this works on older gens that perform a display
reset during gpu reset.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>