kms_rotation_crc: Make tests pass.
Some tests were failing to set size after rotation, which resulted in
some tests failing.

Another reason was that the test assumed that vdisplay < hdisplay, this
is not the case for a 4k DP-MST display, which runs in 1920x2160 mode.

The vtcon switching test was also bogus and caused failures. It's not
always true and even on a system with fbcon I was seeing a test failure
because the vt never switched.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>