lib/igt_kmod: Compare module names with strcmp
The function igt_kmod_is_loaded() returns the wrong value when there is
a module loaded whose name is a prefix of the name supplied as a
parameter. For instance, if the "snd" module is loaded,
igt_kmod_is_loaded("snd_hda_intel") will return true even if that module
isn't loaded, thus causing drv_module_reload to fail in that scenario.

Signed-off-by: Ander Conselvan de Oliveira <ander.conselvan.de.oliveira@intel.com>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>