lib: Add basic support for valgrind annotations.
SIGRTMAX appears to be used by valgrind now for its internal tracking,
so avoid it in the helpers.

Also add some valgrind annotations in gem_mmap, to make sure that its
accesses are tracked correctly. I've also added gem_munmap, but there
are a lot of places that don't use it yet in tests/.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>