tests/kms_cursor_legacy: make_busy() before get_vblank()
VLV is a sloth and so doing anything extra between the first and last
get_vblank() is likely to increase the chance of failing the test.
Let's move the make_busy() stuff to happen before the first get_vblank()
to reduce the amount of work done there.

These tests are still failing quite often on my VLV, but that is mostly
caused by igt_kms's tendency to still re-read all the connector properties
and whatnot as part of the commit operation. I suspect that fixing that
would eliminate most of the spurious failures.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>