iget/gem_exec_reloc: Fix simulated relocations
The contract with the kernel is that the presumed_offset matches the
value written into the batch. In the case where we were creating a new
object to simulate the old being relocation, we were writing some other
value into the batch. It just happens that using GGTT read back on !llc
was causing the original batch to migrated into the aperture, leaving a
hole suitable for the new batch, and the kernel could therefore skip the
relocation (causing us to complain).

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=100674
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>