gem_wsim: Fix master workload handling and stats
Handling was broken in quiet mode which got expose by the
conversion to verbosity.

And I want to have the runtime stats for the master workload
printed in the default verbosity as well.

Also tidy help text adding rtavg to id and group cmd line
options not taking any parameters together.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>