wsim: Limit the information updated during the heartbeat
The heartbeat should be sent after a period of idleness, it's sole
purpose is to update our idea of the external load. As we are sending it
currently in the middle of a busy period, we can reduce it to just
bumping the queue-depth and reusing the most recent clocks, so that we
do not perturb the RT timings greatly.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>