gem_wsim: Add append workload
Used with the '-a' command line switch which follows the same
usage as '-w' and '-W', it enables to add append workload steps
to the end of all normal workloads.

This for example allows running any workload in the real-time
mode:

  gem_wsim -w <some-workload> -a p.16667

Makes a workload to be run with the 60 Hz period.

At the same time fix the periodic mode execution with dropped
frames, or almost dropped frames.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>