gem_wsim: Add global balancing mode
In this mode ('-G' on the command line) all balancing operations
are routed via the first client so the complete balancing state
is shared. In other words the overall balancing behaviours is
like there is only one client submitting the aggregate workload.

This can help with the observed metrics and lead to better
balancing decisions in a lot of cases.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>