igt/gem_shrink: Exercise concurrent calls to i915_gem_shrink()
Just recently I once again made the mistake of thinking we could do a
plain mutex_lock() inside i915_gem_shrink(). However, such a lock is
liable to cyclic deadlocks between multiple relcaimers. This can be
reported by lockdep, but we need contention in the shrinker for it to
spot this particular mistake. The easiest way to explicit cause
contention is via concurrent calls to debugfs/i915_drop_caches whilst
the GPU is busy.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>