gem_wsim: Add static context balancer
This one assigns VCS engines for each created context in a round-
robin fashion. By default each client starts the assignment from
the first engine.

In global mode the engines are assigned in the round-robin fashion
between all the clients - which is identical to the i915 behaviour.
Important difference is that doing it from gem_wsim means the
assignment always start from a known point so the run-to-run
behaviour is identical and repeatable, contrary to when i915 round-
robin assignment is used.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>