igt/vc4_tiling: Test vc4's new set/get_tiling ioctls.
This just checks that the appropriate errors get thrown, and that the
modifier can be set/get successfully, and that the modifier doesn't
leak to other BO allocations.  Testing of scanout will be done with
the writeback support that Boris is building.  The modifier has no
effect on V3D rendering, so no need to test that.

v2: Add to the meson build.

Signed-off-by: Eric Anholt <eric@anholt.net>