igt/gem_exec_fence: Limit history size for !execlists
Only with execlists, is each context using a separate ring buffer and we
can accumulate as many fences as contexts we can allocate. On the hand,
systems with just a single ring buffer for all contexts, we can only
allocate enough contexts/fences as to fill the ring. If we try to add to
many we end up stuck waiting for space.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>