kms_rotation_crc: Fix rotation tests on older platforms again.
I've erroneously assumed the rotation tests only work on gen9+,
and didn't test what would happen on older platforms.

Add an extra testcase that tries to fill the whole screen, and only
run that test on gen8 and lower for the primary plane.

This fixes the testcases failing since the following commit:

commit a7fe05b0a784a7a7c58d36842426579a6e00f5a0
Author: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Date:   Tue Jun 6 16:41:12 2017 +0200

    tests/kms_rotation_crc: Implement stricter rotation tests

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>