tests/chamelium: Set output pipe after mode change to refresh output
This adds a call to igt_output_set_pipe in orde to refresh output via
igt_output_refresh and ensure that mode override can take effect.

Without this change, using a lower resolution during frame dumps
series with mode changes (e.g. test_display_frame_dump) would not
commit the mode change to the output and leave it with a framebuffer
too small for the previously-commited resolution.

This would cause the test to fail. This allows it to pass.

Signed-off-by: Paul Kocialkowski <paul.kocialkowski@linux.intel.com>
Reviewed-by: Lyude Paul <lyude@redhat.com>