igt/gem_mmap_offset: Simulate gdb inspecting any mmap using ptrace()
gdb uses ptrace() to peek and poke bytes of the target's address space.
The kernel must implement an vm_ops->access() handler or else gdb will
be unable to inspect the pointer and report it as out-of-bounds. Worse
than useless as it causes immediate suspicion of the valid GPU pointer.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Mika Kuoppala <mika.kuoppala@linux.intel.com>