configure.ac: Make building chamelium an option
Originally when chamelium support got added we were expecting basically
any distribution to already have the required dependencies for it,
however seeing as someone made chamelium compilation conditional that
must not actually be the case.

This being said however, when compilation of the chamelium tests/libs
was made to only build when all of the dependencies were found, nothing
was actually added to configure.ac to give even the slightest indication
that chamelium support wasn't going to get built. This ended up breaking
my autocompletion for the chamelium source files, leading to a multihour
troubleshooting session where I eventually realized that someone had
changed building chamelium support from mandatory to automatic.

For now, I'll make this conditional so that people who can't satisfy
these dependencies can disable this and everyone else can get warnings
to let them know that they're missing dependencies required to build
this.

Please add messages in the future to our build scripts when these things
get changed :\.

Additionally, we also add xmlrpc_util and xmlrpc_client to the
chamelium dependencies. This isn't required older versions of libxmlrpc,
but it seems that in newer versions they've changed which packages
provide which symbols. It shouldn't break building with older versions
of xmlrpc though.

Signed-off-by: Lyude <lyude@redhat.com>