README: Add information about chamelium dependencies
This adds a list of dependencies required to build chamelium support,
so that what needs to be installed to get it going is more obvious.

As done previously in the file, the list is relevant for Debian and its
derivatives but should provide a template for other systems as well.

Signed-off-by: Paul Kocialkowski <paul.kocialkowski@linux.intel.com>
Reviewed-by: Lyude <lyude@redhat.com>