configure.ac: Disable chamelium by default and add enable argument
Since the chamelium is not a very usual piece of hardware and requires
pulling-in lots of specific dependencies, it makes sense to keep it
disabled by default.

An explicit --enable-chamelium argument is provided to enable it when
necessary. This also leads to more predictable results than
automatically enabling it when its dependencies are met.

Finally, whether chamelium support is enabled is printed by the
information summary at end of the configure run.

Signed-off-by: Paul Kocialkowski <paul.kocialkowski@linux.intel.com>
Reviewed-by: Lyude <lyude@redhat.com>