tests/kms_busy: Only test against one engine
Back when we used cs flips it made sense to go through different
engines, since a buffer busy on an engine that we couldnt' use for
cs flipping ended up in different paths.

But with atomic we use a worker for all flips, and going through the
combinatorial growth of engines just wastes precious machine time.
More so the more modern the platform is.

Of course gem tests should still do some diagonal testing across all
engines, but the kms side can afford to be a bit cheaper.

v2: Make it compile. Oops. Also make sure we don't break
fast-feedback.testlist.

Adding a mode where we depend upon all engines isn't really possible,
because there's only 1 exclusive fence and that's the only thing a
flip waits for.

Cc: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Cc: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Acked-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>