tests/kms_properties: Don't set immutable properties
If the kernel tells us it's immutable, trying to set it probably isn't
going to succeed. There was previously a hard-coded list of immutable
properties, keyed by name, but we can just test the flag instead, which
is more future-proof.

Fixes a failure seen with the IN_FORMATS property.

v2: Remove the entire list apart from DPMS, replacing it with a test for
    immutable.

Signed-off-by: Daniel Stone <daniels@collabora.com>
Reviewed-by: Daniel Vetter <daniel.vetter@intel.com>
Cc: Ben Widawsky <ben@bwidawsk.net>