tests/kms_flip: fix spin_batch conversion
The goal of these subtests was just to delay a kms operation a little
bit. The goal wasn't to
- spin until the operation was completed. That results in a gpu hang,
  and gpu hangs need igt_hang for safety.
- complete it before the operation. That's just pointless.

Fix it by using the timeout support.

This was all broken in the initial conversation:

commit 96ec8cb3b5ec1fc2927d6cff8e09930082301d7e
Author: Abdiel Janulgue <abdiel.janulgue@linux.intel.com>
Date:   Sat Oct 29 01:01:05 2016 +0300

    igt/kms_flip: Use new igt_spin_batch

Note that part of the damage was removed already in

commit b0081ea9cb7d39234fd0fcc64dd56ed4f5d50b05
Author: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Date:   Wed Aug 9 11:11:52 2017 +0200

    tests/kms_flip: Remove $engine-flip-vs-dpms/modeset

But this patch also made things worse, because it remove code that
wasn't dead yet (because it didn't remove all the dpms vs spinning
batch tests).

v2: Augment commit message and use macro for 1s-in-ns.

Cc: Tomeu Vizoso <tomeu.vizoso@collabora.com>
Cc: Abdiel Janulgue <abdiel.janulgue@linux.intel.com>
Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>